break: move utility functions and classes into utils directory
